Rename package-level ESLint configs to prevent auto-discovery
ESLint v10 searches for eslint.config.* files in subdirectories and
uses them as standalone configs, bypassing the root config. Rename
the package-level override files so they are only loaded via explicit
import from the root eslint.config.mjs.
