+++ packages/wp-build/eslint-overrides.cjs
